% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_colour_wolves}
\alias{scale_colour_wolves}
\title{Colour scale constructor for wwfc colours}
\usage{
scale_colour_wolves(
  palette = "wooly_bully",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in wwfc_palettes}

\item{discrete}{Boolean indicating whether colour aesthetic is discrete or
not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_color_gradientn() according to whether discrete arg is TRUE or FALSE}
}
\description{
Custom colour scale functions for ggplot2. Handles discrete or continuous
scales.
}
\examples{
# Color by discrete variable using default palette
library(ggplot2)
ggplot(iris, aes(Sepal.Width, Sepal.Length, colour = Species)) +
  geom_point(size = 4) +
  wolves_theme() +
  scale_colour_wolves()

# Color by numeric variable with likert palette
ggplot(iris, aes(Sepal.Width, Sepal.Length, color = Sepal.Length)) +
  geom_point(size = 4) +
  wolves_theme() +
  scale_colour_wolves(discrete = FALSE, palette = "sir_jack")
}
