% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{scale_fill_wolves}
\alias{scale_fill_wolves}
\title{Fill scale constructor for wwfc colours}
\usage{
scale_fill_wolves(
  palette = "wooly_bully",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in wwfc_palettes}

\item{discrete}{Boolean indicating whether fill aesthetic is discrete or not}

\item{reverse}{Boolean indicating whether the palette should be reversed}

\item{...}{Additional arguments passed to discrete_scale() or
scale_fill_gradientn() according to whether discrete arg is TRUE or FALSE}
}
\description{
Custom fill scale functions for ggplot2. Handles discrete or continuous
scales.
}
\examples{
# Fill by discrete variable with different palette + remove legend (guide)
# Note: you will probably want to avoid using colour in this way!
library(ggplot2)
ggplot(mpg, aes(manufacturer, fill = manufacturer)) +
  geom_bar() +
  wolves_theme(grid = "vgrid") +
  coord_flip() +
  scale_y_continuous(expand = expansion(mult = c(0, 0.05))) +
  scale_fill_wolves(palette = "wooly_bully", guide = "none") +
  labs(title = "scale_fill_wolves() can generate many colours from a palette",
       subtitle = "Using colour in this way (despite consistent data type) is not recommended!")
}
