% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{view_palette}
\alias{view_palette}
\title{View a palette}
\usage{
view_palette(palette)
}
\arguments{
\item{palette}{A character vector of colours to display.}
}
\description{
Displays a palette using geom_col() and theme_void().
}
\examples{
view_palette(wwfc_pal("wooly_bully")(8))

# display wwfc palettes
library(ggplot2)

wooly_bully <- view_palette(wwfc_pal("wooly_bully")(5)) +
    labs(title = '"wooly_bully" palette') +
    theme(plot.title = element_text(size = 18, hjust = .5))
sir_jack <- view_palette(wwfc_pal("sir_jack")(5)) +
  labs(title = '"sir_jack" palette') +
  theme(plot.title = element_text(size = 18, hjust = .5))
pal_plot <- cowplot::plot_grid(
  wooly_bully, sir_jack,
  nrow = 1
  )
pal_plot
}
