% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wolves_theme.R
\name{wolves_theme}
\alias{wolves_theme}
\title{Theme plots in Wolverhampton Wanderers style}
\usage{
wolves_theme(
  media = "web",
  grid = "hgrid",
  panel_border = FALSE,
  base_size = 11
)
}
\arguments{
\item{media}{A string, either "print" or "web". Should the font size
be optimised for print or web outputs? Default: web.}

\item{grid}{A string, either "hgrid", "vgrid" or "fullgrid". Allows for
setting optimal orientation of gridlines and easy adjustment after using
\code{coord_flip()}. See examples. Default: hgrid.}

\item{panel_border}{logical, should a panel border be drawn? Intended for use
with faceted plots where a panel border may be needed if there are many
panels.}

\item{base_size}{integer. Base font size, given in pts. Default: 11.}
}
\description{
ggplot2 minimal theme which optimises font size (for web or print) and
the use/orientation of gridlines (major gridlines only).
}
\details{
Uses the ggplot2 \code{theme} to set consistent plot theme elements.
Overwrites any preceding theme calls and can be overwritten by subsequent
ones.
}
\examples{
suppressMessages(library(dplyr))
library(ggplot2)

# wolves theme with font size optimised for web outputs
ggplot(iris, aes(x = Species, y = Sepal.Length)) +
  geom_boxplot(fill = wwfc_cols("green")) +
  wolves_theme(media = "web") +
  labs(title = "Virginica sepals are longest",
       subtitle = "Sepal length (mm) by species",
       y = "")

# wolves theme with font size optimised for web outputs
ggplot(iris, aes(x = Species, y = Sepal.Length)) +
  geom_boxplot(fill = wwfc_cols("green")) +
  wolves_theme(media = "print") +
  labs(title = "Virginica sepals are longest",
       subtitle = "Sepal length (mm) by species",
       y = "")

# density plot in Wolves style
ggplot(iris, aes(x = Sepal.Length, fill = Species)) +
  geom_density(alpha = .7) +
  wolves_theme() +
  scale_y_continuous(expand = expansion(mult = c(0, .05))) +
  labs(
    title = "Setosa sepals may be reliably distinguished by length",
    subtitle = "Distribution of sepal length (mm) by species",
    y = "",
    fill = ""
    ) +
  theme(plot.title.position = "plot") +
  scale_fill_wolves()

# horizontal gridlines
iris \%>\%
  group_by(Species) \%>\%
  summarise(mean = mean(Sepal.Length), .groups = "drop") \%>\%
  ggplot(aes(x = Species, y = mean)) +
  geom_col(fill = wwfc_cols("green")) +
  wolves_theme() +
  scale_y_continuous(
    limits = c(0,7),
    breaks = c(0:8),
    expand = expansion(mult = c(0, .05))
    ) +
  labs(title = "Virginica sepals are longest",
       subtitle = "Sepal length (mm) by species",
       y = "")

# vertical gridlines
iris \%>\%
  group_by(Species) \%>\%
  summarise(mean = mean(Sepal.Length), .groups = "drop") \%>\%
  ggplot(aes(x = Species, y = mean)) +
  geom_col(fill = wwfc_cols("green")) +
  coord_flip() +
  wolves_theme(grid = "vgrid") +
  scale_y_continuous(
    limits = c(0,7),
    breaks = c(0:8),
    expand = expansion(mult = c(0, .05))
    ) +
  labs(title = "Virginica sepals are longest",
       subtitle = "Sepal length (mm) by species",
       y = "")

# full grid with horizontal and vertical gridlines
iris \%>\%
  ggplot(aes(x = Sepal.Width, y = Sepal.Length, colour = Species)) +
  geom_point(size = 3) +
  wolves_theme(grid = "fullgrid") +
  scale_colour_wolves() +
  scale_x_continuous(expand = expansion(mult = c(0.05, .05))) +
  labs(title = "Different strategies to achieve sepal area",
       subtitle = "Sepal dimensions (mm) by species",
       y = "Length",
       x = "Width",
       colour = "") +
  theme(legend.position = "right")

}
