% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{wwfc_cols}
\alias{wwfc_cols}
\title{Function to access wwfc colours}
\usage{
wwfc_cols(...)
}
\arguments{
\item{...}{Character names of wwfc_colours}
}
\description{
Allows you to access wwfc brand colours in a robust and flexible way.
Supply colour names and the function ouputs hex codes. An empty function call
returns a named vector of all the colours used in available wwfc palettes.
}
