% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palettes.R
\name{wwfc_pal}
\alias{wwfc_pal}
\title{Return a function to interpolate an wwfc colour palette}
\usage{
wwfc_pal(palette = "wooly_bully", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in wwfc_palettes}

\item{reverse}{logical indicating whether the palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\description{
A function that is mainly intended for internal package use. Gets a pallete
by name from a list. Uses a boolean condition determining whether to reverse
the order or not, and additional arguments to pass on to colorRampPallete()
(such as an alpha value).
}
\details{
Returns another function, which will interpolate the palette colors for a
certain number of levels, making it possible to create shades between our
original colours.

The complete list of options is "wooly_bully" (the default) & "sir_jack". Use
e.g. view_palette("wooly_bully") to see the 5 core colours that make up a
palette.
}
\examples{
wwfc_pal("wooly_bully")(8)
wwfc_pal("sir_jack")(12)
}
